pyproject.toml
[tool.poetry]
name = "langgraph_react"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
python-dotenv = "^1.0.1"
langchain = "^0.2.1"
langchain-openai = "^0.1.7"
langgraph = "^0.0.55"
langchainhub = "^0.1.15"
black = "^24.4.2"
isort = "^5.13.2"
tavily-python = "^0.3.3"
langchain-community = "^0.2.1"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
